requirements.txt
attrs==19.3.0
azure-common==1.1.25
azure-core==1.4.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.3.0
boxsdk==2.7.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
idna==2.8
importlib-metadata==1.5.0
isodate==0.6.0
more-itertools==8.2.0
msrest==0.6.13
oauthlib==3.1.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.12.0
zipp==3.0.0
